requirements.txt
folium==0.18.0
haversine==2.8.1
markdown-it-py==3.0.0
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
plotly-express==0.4.1
streamlit==1.40.2
streamlit_folium==0.23.2